% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmodelfuncs.R
\name{constCycle}
\alias{constCycle}
\title{Constant cycle component}
\usage{
constCycle(parmlist)
}
\arguments{
\item{parmlist}{A list of parameters as generated by \code{\link{parmList}}.}
}
\value{
A vector \code{ct} of the same length as the data \code{yt}.
}
\description{
Constant cycle component
}
\details{
This function generates state space model matrices related to an a
  constant cyclic component. Such a specification is usually more slowly
  varying than a seasonal component, say for a yearly cycle with daily data.
  The model has the form: \deqn{y(t) = c(t) + extra stuff} where \deqn{c(t) =
  c_1 * cos(lam*t) + c_2 * sin(lam*t)} with \eqn{lam=2\pi/period} and \eqn{t}
  simply being the times at which the observations occurred.

  Note that this really is no more than a time varying intercept in the
  observation equation. For a particular period (\code{parmlist$period}),
  this function produces the appropriate model matrices above.
}
