% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmodelfuncs.R
\name{generateSSmodel}
\alias{generateSSmodel}
\title{Generate (univariate) state space model matrices.}
\usage{
generateSSmodel(trendModel, seasonalModel = "none", cycleModel = "none",
  parmlist = parmList())
}
\arguments{
\item{trendModel}{The name of a specific trend type. Options are
\code{LocalLevel}, \code{LocalLinTrend}, \code{SmoothSlope},
\code{ConstantSlope}, and \code{IntRandWalk}. Input should be a quoted
string. This updates the parmlist to reflect the specific trend choice. Any
other string simply applies \code{\link{trendModelArb}} to the parmlist.}

\item{seasonalModel}{The name of a seasonal type, either \code{trigSeasonal}
or \code{unstrucSeasonal}. The default is \code{'none'}.}

\item{cycleModel}{The name of a cycle type, either \code{tvarCycle} or
\code{constCycle}. The default is \code{'none'}.}

\item{parmlist}{A list of parameters. Likely the output of the function
\code{\link{parmList}}. These will be used to parameterize the model
matrices.}
}
\value{
A list of components

  \code{dt}	A matrix giving the intercept of the transition equation.\cr
  \code{ct}	A matrix giving the intercept of the measurement equation.\cr
  \code{Tt}	An array giving the factor of the transition equation.\cr
  \code{Zt}	An array giving the factor of the measurement equation.\cr
  \code{HHt}	An array giving the variance of the innovations of the
  transition equation.\cr \code{GGt}	An array giving the variance of the
  disturbances of the measurement equation.\cr \code{parmlist} The parameters
  used to generate the model (may modify the input).
}
\description{
This function generates all necessary matrices for various types of
univariate structural state space models. Not particularly useful on its own,
it serves as input for the \code{\link{fkf}} function.
}
\examples{
ssMod = generateSSmodel('LocalLinTrend',cycleModel = 'tvarCycle',
   parmlist=parmList(sig2xi=.5,sig2eta=.1))
simOut = simSS(250, ssMod)
plot(1:250,simOut$y,ty='l',lwd=2)
## Plot the trend
lines(1:250,c(1,1)\%*\%simOut$a[1:2,],col=2)
lines(1:250,c(1,1)\%*\%simOut$a[3:4,]+ #the cycle added to the trend
   c(1,1)\%*\%simOut$a[1:2,],col=4)
}
