% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmodelfuncs.R
\name{parmList}
\alias{parmList}
\title{Generate parameters for various univariate state-space models.}
\usage{
parmList(...)
}
\arguments{
\item{...}{Named parameters for univariate state space models. These can be
  any of the following (defaults given)

  \code{sig2epsilon = 1, sig2xi = 1, sig2eta = 1, sig2omega = 1, sig2cycle =
  1,d = 2, r = 2, constSlope = 0, nu = FALSE, phi1 = 1, phi2 = 1, period =
  12, damp = 1, ccycle = double(2), obsConst = 0, timepoints = 1:100)}

  Generally, these are set by other functions (\code{\link{trendModelArb},
  \link{unstrucSeasonal}, \link{trigSeasonal}, \link{tvarCycle},
  \link{constCycle}, or \link{generateSSmodel}}.}
}
\value{
A list of parameters for use in generating certain kinds of
  univariate state space models.
}
\description{
Generate parameters for various univariate state-space models.
}
