% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fkf2.R
\name{simSS}
\alias{simSS}
\title{Simulate linear state space models}
\usage{
simSS(nobs, modelMats, Tt = matrix(1), Zt = matrix(1), HHt = diag(1,
  dim(Tt)[1]), GGt = diag(1, nrow(Zt), ncol(Zt)), dt = double(dim(Tt)[1]),
  ct = double(dim(Zt)[1]))
}
\arguments{
\item{nobs}{The desired number of timepoints.}

\item{modelMats}{A list of the matrices below, likely as output from the
function \code{\link{generateSSmodel}}}

\item{Tt}{A matrix giving the factor of the transition equation.}

\item{Zt}{A matrix giving the factor of the measurement equation.}

\item{HHt}{A matrix giving the variance of the innovations of the transition
equation.}

\item{GGt}{A matrix giving the variance of the disturbances of the
measurement equation.}

\item{dt}{A matrix giving the intercept of the transition equation.}

\item{ct}{A matrix giving the intercept of the measurement equation.}
}
\value{
A list with the following components:

  \code{a} a matrix of the simulated states of dimension \eqn{m x nsimul}\cr
  \code{y} a matrix of the simulated observations of dimension \eqn{d x
  nsimul}
}
\description{
This function simulates a time-invariant state space model. That is, the
parameters are constant in time. The state space model is represented by the
transition equation and the measurement equation. Let \code{m} be the
dimension of the state variable, \code{d} be the dimension of the
observations, and \code{n} the number of observations. The transition
equation and the measurement equation are given by \deqn{a(t + 1) = d(t) +
T(t) a(t) + H(t) \eta(t)} \deqn{y(t) = c(t) + Z(t) a(t) + G(t) \epsilon(t),}
where \eqn{\eta(t)} and \eqn{\epsilon_t} are iid \eqn{N(0, I_m)} and iid
\eqn{N(0, I_d)}, respectively, and \eqn{\alpha(t)} denotes the state
variable. The parameters admit the following dimensions:
}
\examples{

Tt = matrix(c(.9,.1,.2,.7),2)
Zt = matrix(c(1,0,.2,0,1,.8),3)
HHt = .5*diag(2)
GGt = .1*diag(3)
outSim = simSS(100, Tt=Tt, Zt=Zt, HHt=HHt, GGt=GGt)
}
