% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmodelfuncs.R
\name{trendModelArb}
\alias{trendModelArb}
\title{Univariate trend components}
\usage{
trendModelArb(parmlist)
}
\arguments{
\item{parmlist}{The input contains named elements \code{d, phi1, r, phi2, sig2eta, sig2xi}.
  All are scalars parameterizing the above model. It also contains a logical
  for \code{nu}, which determines whether to include the \eqn{nu} component
  (default is \code{FALSE}). All of these can be set with
  \code{\link{parmList}}.

  The default values (generated by \code{\link{parmList}}) are

  \code{d} = 2 \cr \code{phi1} = 1\cr \code{r} = 2\cr \code{phi2} = 1\cr
  \code{sig2eta} = 1\cr \code{sig2xi} = 1\cr \code{nu} = \code{FALSE}}
}
\value{
A list of matrices with components \code{Z}, \code{Tt}, \code{Q}, and
  \code{R}.
}
\description{
This function can generate the appropriate state and transition matrices
(\eqn{Tt}, \eqn{Zt}, \eqn{GGt}, \eqn{HHt}) for different possible trend
components. Possible univariate trends are given with: \deqn{y(t) = mu(t) +
extra stuff} \deqn{(1 - phi1 L)^d mu(t) = nu(t-1) + xi(t-1)} \deqn{(1 - phi2
L)^r nu(t) = eta(t)} Noise terms \eqn{eta} and \eqn{xi} are iid normal mean 0
with variances \code{sig2eta} etc. This function generates all possible
combinations of these. See Chapter 3 in (Durbin and Koopman) for more
information. "Extra stuff" can be simply an error variance or may include
seasonal, cyclical, or other components.
}
