% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmodelfuncs.R
\name{tvarCycle}
\alias{tvarCycle}
\title{Time-varying cycle component}
\usage{
tvarCycle(parmlist)
}
\arguments{
\item{parmlist}{A list of parameters as generated by \code{\link{parmList}}.}
}
\value{
A list of matrices with components \code{Z}, \code{Tt}, \code{Q}, and
  \code{R}.
}
\description{
Time-varying cycle component
}
\details{
This function generates state space model matrices related to an a
  time-varying cyclic component. Such a specification is usually more slowly
  varying than a seasonal component, say for a yearly cycle with daily data.
  The model has the form: \deqn{y(t) = Z_c * c(t) + extra stuff} \deqn{c(t+1)
  = Tt_c * c(t) + R * sig_cycle}

  In particular, the vector c(t) evolves as \deqn{c_1(t+1) = c_1(t) *
  cos(lam) + c_2(t) * sin(lam) + omega_1} \deqn{c_2(t+1) = -c_1(t) * sin(lam)
  + c_2(t) * cos(lam) + omega_2} where \eqn{omega_1} and \eqn{omega_2} are
  iid normal with mean 0 and variance sig2cycle, and \eqn{lam = 2\pi/period}.

  For a particular period (\code{parmlist$period}), this function produces
  the appropriate model matrices above.
}
