% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmodelfuncs.R
\name{unstrucSeasonal}
\alias{unstrucSeasonal}
\title{Unstructured seasonal component}
\usage{
unstrucSeasonal(parmlist)
}
\arguments{
\item{parmlist}{A list of parameters as generated by \code{\link{parmList}}.}
}
\value{
A list of matrices with components \code{Z}, \code{Tt}, \code{Q}, and
  \code{R}.
}
\description{
Unstructured seasonal component
}
\details{
This function generates state space model matrices related to an an
  unstructured seasonal component. The model has the form: \deqn{y(t) = Z_gam
  * gam(t) + extra stuff} \deqn{gam(t+1) = Tt_gam * gam(t) + R * sig_omega}

  For a particular period (\code{parmlist$period}), this function produces
  the appropriate model matrices above. Note that the size of the state
  vector (related to the seasonal) is equal to the period. This can get
  large.
}
